+++ src/main/java/cc/polyfrost/oneconfig/gui/elements/config/ConfigSwitch.java
@@ -88,7 +88,7 @@ public class ConfigSwitch extends BasicOption {
     }
 
     @Override
-    protected int getNameX(int x) {
+    protected float getNameX(int x) {
         return x + 50;
     }
